Added support for keyboard, big user screen, proper error messages and fixed bug
This commit adds support for a keyboard at $9000 (the emulator places a keycode in $9000 if the word there is 0 or beeps if it is not - the program has
to get the code and set it to 0).

Also a big user screen has been added which can be used to view the programs in a larger area and accept input.

Finally a bug in the emulator has been fixed. The bug caused PUSH/POP values to be executed even if the "not execute next" flag was set.